### 2025-04-06 #8011 @editorialbot
[PRE REVIEW]: Ziplign: a simple-to-use interactive tool to compare bacterial genomes
**Submitting author:** @martinghunt (<a href="http://orcid.org/0000-0002-8060-4335">Martin Hunt</a>)
**Repository:** https://github.com/martinghunt/ziplign
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.1
