### 2025-05-17 #8011 @Kevin-Mattheus-Moerman
@martinghunt - thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editors most appropriate for your area are all already rather busy.
