+++ docs/installation.rst
@@ -1,5 +1,5 @@
-Installing TNA
-==============
+Installing Ziplign
+==================
 
 Supported Operating Systems
 ---------------------------
@@ -9,15 +9,15 @@ Supported Operating Systems
 * Linux (tested on Ubuntu 20.04/x86, 22.04/x86, and 24.04/aarch64)
 
 
-Download TNA
-------------
+Download Ziplign
+----------------
 
-Go to the `TNA release page <https://github.com/martinghunt/tna/releases>`_
+Go to the `Ziplign release page <https://github.com/martinghunt/ziplign/releases>`_
 and download the correct version for your computer:
 
-* Windows 11 - ``tna.vX.Y.Z.windows.exe``
-* macOS - ``tna.vX.Y.Z.mac.dmg``
-* Linux x86_64/arm64 - ``tna.vX.Y.Z.linux.x86_64``/``tna.vX.Y.Z.linux.arm64``
+* Windows 11 - ``ziplign.vX.Y.Z.windows.exe``
+* macOS - ``ziplign.vX.Y.Z.mac.dmg``
+* Linux x86_64/arm64 - ``ziplign.vX.Y.Z.linux.x86_64``/``ziplign.vX.Y.Z.linux.arm64``
 
 The actual filenames have the release number in them instead of ``X.Y.Z``
 in the above filenames.
@@ -29,13 +29,13 @@ Windows
 
 Double-click on the downloaded file and it should just work.
 You may find Windows Defender popping up, in which case you will need to
-tell it to allow TNA to run.
+tell it to allow Ziplign to run.
 
 
 macOS
 ^^^^^
 
-Double-click the downloaded ``tna.mac.dmg`` file. It contains the app ``tna.app``.
+Double-click the downloaded ``ziplign.mac.dmg`` file. It contains the app ``ziplign.app``.
 Drag or copy it into your Applications folder (or wherever you like depending on
 how you organise your files).
 
@@ -43,11 +43,11 @@ macOS Gatekeeper will probably block it from running. If this happens,
 you have two options:
 
 1. Go to "Privacy & Security" in the Settings app. Scroll to the bottom
-   and TNA should be there for you to allow it
+   and Ziplign should be there for you to allow it
 2. In a terminal, run this command:
-   ``xattr -d com.apple.quarantine -r tna.app``.
+   ``xattr -d com.apple.quarantine -r ziplign.app``.
 
-Then TNA should just work. We apologise for the inconvenience, but this
+Then Ziplign should just work. We apologise for the inconvenience, but this
 process is standard for apps that have not been
 `"notarized" by Apple <https://www.youtube.com/watch?v=X6HZlpPGFf0>`_ (which
 means paying an annual fee).
@@ -60,16 +60,16 @@ You may need to make the downloaded binary file executable
 a terminal if you want to see logging) should just work.
 
 
-First time running TNA
-----------------------
+First time running Ziplign
+--------------------------
 
-The first time you run TNA, it will automatically download two extra programs:
+The first time you run Ziplign, it will automatically download two extra programs:
 
-1. ``tnahelper``. This is a separate command line program made for TNA, which
+1. ``zlhelper``. This is a separate command line program made for Ziplign, which
    handles bioinformatics tasks (the source code is here:
-   https://github.com/martinghunt/tnahelper). Don't worry, you won't ever
-   have to run it yourself - TNA uses it in the background.
-2. NCBI-blast+. Specifically, TNA gets the programs ``makeblastdb`` and
+   https://github.com/martinghunt/zlhelper). Don't worry, you won't ever
+   have to run it yourself - Ziplign uses it in the background.
+2. NCBI-blast+. Specifically, Ziplign gets the programs ``makeblastdb`` and
[…]
