### 2025-06-23 9dcd8ec
updates after opening in godot ide
