+++ export_presets.cfg
@@ -9,9 +9,11 @@ custom_features=""
 export_filter="all_resources"
 include_filter=""
 exclude_filter=""
-export_path="ziplign.mac.dmg"
+export_path="../../tmp/ziplign.mac.dmg"
+patches=PackedStringArray()
 encryption_include_filters=""
 encryption_exclude_filters=""
+seed=0
 encrypt_pck=false
 encrypt_directory=false
 script_export_mode=2
@@ -32,7 +34,8 @@ application/short_version=""
 application/version=""
 application/copyright=""
 application/copyright_localized={}
-application/min_macos_version="10.12"
+application/min_macos_version_x86_64="10.12"
+application/min_macos_version_arm64="11.00"
 application/export_angle=0
 display/high_res=true
 application/additional_plist_content=""
@@ -70,6 +73,7 @@ codesign/entitlements/app_sandbox/files_music=0
 codesign/entitlements/app_sandbox/files_movies=0
 codesign/entitlements/app_sandbox/files_user_selected=0
 codesign/entitlements/app_sandbox/helper_executables=[]
+codesign/entitlements/additional=""
 codesign/custom_options=PackedStringArray()
 notarization/notarization=0
 privacy/microphone_usage_description=""
@@ -247,6 +251,7 @@ open \"{temp_dir}/{exe_name}.app\" --args {cmd_args}"
 ssh_remote_deploy/cleanup_script="#!/usr/bin/env bash
 kill $(pgrep -x -f \"{temp_dir}/{exe_name}.app/Contents/MacOS/{exe_name} {cmd_args}\")
 rm -rf \"{temp_dir}\""
+application/min_macos_version="10.12"
 
 [preset.1]
 
@@ -260,8 +265,10 @@ export_filter="all_resources"
 include_filter=""
 exclude_filter=""
 export_path="ziplign.linux.arm64"
+patches=PackedStringArray()
 encryption_include_filters=""
 encryption_exclude_filters=""
+seed=0
 encrypt_pck=false
 encrypt_directory=false
 script_export_mode=2
@@ -304,8 +311,10 @@ export_filter="all_resources"
 include_filter=""
 exclude_filter=""
 export_path="ziplign.linux.x86_64"
+patches=PackedStringArray()
 encryption_include_filters=""
 encryption_exclude_filters=""
+seed=0
 encrypt_pck=false
 encrypt_directory=false
 script_export_mode=2
@@ -348,8 +357,10 @@ export_filter="all_resources"
 include_filter=""
 exclude_filter=""
 export_path="ziplign.windows.exe"
+patches=PackedStringArray()
 encryption_include_filters=""
 encryption_exclude_filters=""
+seed=0
 encrypt_pck=false
 encrypt_directory=false
 script_export_mode=2
