### 2025-06-23 650a7e5
Update docs screenshots
