### 2025-06-24 #8011 @martinghunt
@Kevin-Mattheus-Moerman thanks, I've updated the paper and GitHub repo. I'm not sure what needs doing here to reflect the updates.

The tna GitHub repo has been renamed to ziplign, so GitHub will redirect. It's here:
https://github.com/martinghunt/ziplign

It's v1.0.1 of the code instead of v0.3.0. Is the same branch "paper" that has paper.md.

The new title is "Ziplign: a simple-to-use interactive tool to compare bacterial genomes" - ie TNA -> Ziplign.
