### 2025-06-30 #8011 @Kevin-Mattheus-Moerman
@martinghunt I have just triggered a scope review by the editorial board. This is because I need help to determine if this work is in scope for JOSS given its narrow functionality. During the review we'll determine if the work meets our [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion. The review should take about two weeks to complete.
