### 2025-08-24 #8011 @abhishektiwari
@martinghunt Following on my last comment. Let me know if any questions.

> [@martinghunt](https://github.com/martinghunt) I will be helping to edit this paper. Can you please
>
> * add contribution guideline and to mention/link to them in your README. You could for instance create a CONTRIBUTING.md file (see an [example](https://contributing.md/example/) here or you can also use [this generator](https://contributing.md/generator/).
> * can you please clarify the contribution of second author on the paper `Zamin Iqbal`? I do not see any code contribution from `Zamin Iqbal` [here](https://github.com/martinghunt/ziplign/graphs/contributors).
>
> Once you have completed above 2, please feel free to suggest a few potential reviewers. You can also share GitHub handles of potential reviewers. Please be sure not to @-tag your suggestions, to avoid unwanted GitHub notifications to potential reviewers. Please also read the our [conflict of interest](https://github.com/openjournals/joss/blob/main/COI.md) (COI) when recommending reviewers.
