### 2025-09-05 d891a29
Add CONTRIBUTING.md
