### 2025-09-05 #8011 @martinghunt
OK, I've added a CONTRIBUTING file, based on the generator you suggested.

Zamin helped to write the manuscript. Should I add an a author contributions section to make it clearer?
