### 2025-09-14 #8011 @abhishektiwari
👋 @mbhall88 @ammaraziz @vinisalazar @mjsull @warrenlr, would any of you be willing to review this submission for JOSS?  The submission under consideration is "Ziplign: a simple-to-use interactive tool to compare bacterial genomes".

Based on your experience, we think you might be able to provide a great review of this submission. If not, could you recommend any potential reviewers?

Thanks
Abhi
