### 2025-09-14 #8011 @ammaraziz
Happy to review but my GDScript knowledge is zero so I will be mostly testing the GUI.

@abhishektiwari Can you comment on the requirement for testing for a GUI application? I don't see any in the repo (@martinghunt please correct me).
