### 2025-09-15 #8011 @abhishektiwari
Thanks @ammaraziz. Ack. Let me add you add you as reviewer.

I will let @martinghunt  comment on testing. However, you can find installer software [here](https://github.com/martinghunt/ziplign/releases). Once installed UI is launched and then you can follow [instructions as documented](https://ziplign.readthedocs.io/en/v1.0.1/loading.html).

@martinghunt - I think updating the `README.md` on Github will help users on how to quickly get started without clicking through multiple links just to understand how to install it.
