### 2025-09-15 #8011 @ammaraziz
Apologies let me clarify I am referring to CI/CD style testing as [mentioned here](https://joss.readthedocs.io/en/latest/review_criteria.html#tests). If I read correctly, a manual testing procedure that tests the interface (but not directly the code) is acceptable?
