### 2025-09-15 #8011 @abhishektiwari
In this case using UI to perform manual E2E test will be acceptable however I genuinely think software should provide some level of automated tests to improve chances of community’s contributions.
