### 2025-09-17 martinghunt/ziplign#28 @martinghunt
JOSS review: add quick start to readme
See https://github.com/openjournals/joss-reviews/issues/8011#issuecomment-3290125756
