### 2025-09-18 #9004 @abhishektiwari
Comment from @martinghunt  on pre-review thread on automated testing https://github.com/openjournals/joss-reviews/issues/8011#issuecomment-3302438884
