### 2025-10-15 #9004 @ammaraziz
Confirming I've created a few issues in the repo for Martin, some are suggestions (optional) and others are required as per the JOSS rules.
