### 2025-10-23 4662bf6
Add quick start instructions to readme
