### 2025-10-23 d7e48ba
Add manual end-to-end tests; formatting tweaks
