+++ CONTRIBUTING.md
@@ -5,30 +5,30 @@ First off, thanks for taking the time to contribute! ❤️
 
 All types of contributions are encouraged and valued. See the [Table of Contents](#table-of-contents) for different ways to help and details about how this project handles them. Please make sure to read the relevant section before making your contribution. It will make it a lot easier for us maintainers and smooth out the experience for all involved. The community looks forward to your contributions. 🎉
 
-> And if you like the project, but just don't have time to contribute, that's fine. There are other easy ways to support the project and show your appreciation, which we would also be very happy about:
-> - Star the project
-> - Tweet about it
-> - Refer this project in your project's readme
-> - Mention the project at local meetups and tell your friends/colleagues
+And if you like the project, but just don't have time to contribute, that's fine. There are other easy ways to support the project and show your appreciation, which we would also be very happy about:
+ - Star the project
+ - Tweet about it
+ - Refer this project in your project's readme
+ - Mention the project at local meetups and tell your friends/colleagues
 
 <!-- omit in toc -->
 ## Table of Contents
 
 - [I Have a Question](#i-have-a-question)
-  - [I Want To Contribute](#i-want-to-contribute)
-  - [Reporting Bugs](#reporting-bugs)
-  - [Suggesting Enhancements](#suggesting-enhancements)
-  - [Your First Code Contribution](#your-first-code-contribution)
-  - [Improving The Documentation](#improving-the-documentation)
+- [I Want To Contribute](#i-want-to-contribute)
+- [Reporting Bugs](#reporting-bugs)
+- [Suggesting Enhancements](#suggesting-enhancements)
+- [Code Contributions](#code-contributions)
+- [Improving The Documentation](#improving-the-documentation)
 - [Styleguides](#styleguides)
-  - [Commit Messages](#commit-messages)
+- [Commit Messages](#commit-messages)
 - [Join The Project Team](#join-the-project-team)
 
 
 
 ## I Have a Question
 
-> If you want to ask a question, we assume that you have read the available [Documentation](https://ziplign.readthedocs.io).
+If you want to ask a question, we assume that you have read the available [Documentation](https://ziplign.readthedocs.io).
 
 Before you ask a question, it is best to search for existing [Issues](https://github.com/martinghunt/ziplign/issues) that might help you. In case you have found a suitable issue and still need clarification, you can write your question in this issue. It is also advisable to search the internet for answers first.
 
@@ -77,7 +77,7 @@ Once it's filed:
 
 - The project team will label the issue accordingly.
 - A team member will try to reproduce the issue with your provided steps. If there are no reproduction steps or no obvious way to reproduce the issue, the team will ask you for those steps and mark the issue as `needs-repro`. Bugs with the `needs-repro` tag will not be addressed until they are reproduced.
-- If the team is able to reproduce the issue, it will be marked `needs-fix`, as well as possibly other tags (such as `critical`), and the issue will be left to be [implemented by someone](#your-first-code-contribution).
+- If the team is able to reproduce the issue, it will be marked `needs-fix`, as well as possibly other tags (such as `critical`), and the issue will be left to be [implemented by someone](#code-contributions).
 
 
 ### Suggesting Enhancements
@@ -104,14 +104,52 @@ Enhancement suggestions are tracked as [GitHub issues](https://github.com/martin
 - **Explain why this enhancement would be useful** to most Ziplign users. You may also want to point out the other projects that solved it better and which could serve as inspiration.
 
 
-### Your First Code Contribution
+### Code Contributions
 
 Ziplign is developed using the Godot engine and IDE available for free from here: https://godotengine.org/download/.
 It is available for Windows, MacOS and Linux.
 Please use the Godot IDE it to modify the code, and let it do the formatting (indendation etc).
[…]
