### 2025-10-27 3d0a872
Improve wording for mauve test contigs
See JOSS paper issue #23
