### 2025-11-11 8308315
gif 80% width instead of 60%
