### 2025-11-24 646329e
Improve mauve wording; JOSS issue #38
