### 2025-12-14 #9004 @abhishektiwari
@martinghunt

Please see following corrections/recommendation on paper

Line 12: It should be "genome comparison" (no need to hyphenate)
Line 15, 49: It should be BLAST+ not blast+
Line 17: Reads "Figure 1: Figure 1 Screenshots of Ziplign...", should be "Figure 1: Screenshots of Ziplign..."
Line 48, 49: It should be BLASTN not blastn

Missing spaces before citations:
Line 15: "blast+(Camacho et al., 2009)" should be "BLAST+ (Camacho et al., 2009)"
Line 27: "ACT(T. J. Carver et al., 2005)" should be "ACT (T. J. Carver et al., 2005)"
Line 27: "Mauve(Darling et al., 2004)" should be "Mauve (Darling et al., 2004)"
Line 29: "Artemis(T. Carver et al., 2012)" should be "Artemis (T. Carver et al., 2012)"

Line 37: Should be "A, C, G, T" (spaces after commas)
Line 46: Should be "FASTA, FASTQ, GFF3, EMBL, and GenBank"
Line 61-62: Suggested edit "free, open-source, MIT-licensed game engine Godot"

Incorrectly formatted reference titles in rendered PDF, correct titles as original articles. I checked you `bib` file and they look correct but some reason PDF references are incorrectly formatted.

- BLAST+: architecture and applications
- ACT: the Artemis comparison tool
- Artemis: an integrated platform for visualization and analysis of high-throughput sequence-based experimental data
- Escherichia coli K-12: a cooperatively developed annotation snapshot—2005

Please let me know once updated, and I will review one more time before making recommendations. We nearly there.
