### 2025-12-17 #9004 @abhishektiwari
@martinghunt - I hope you saw my last message. Please note JOSS converts reference titles to sentence case and to preserve the case use `{}`.
