### 2025-12-17 44dd4c2
Corrections for editor
