### 2025-12-17 e78dbb4
fix capitalisation in ref titles
