+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{act,
-	title = {{ACT}: the {Artemis} comparison tool},
+	title = {{ACT}: {the} {Artemis} comparison tool},
 	volume = {21},
 	issn = {1367-4803, 1460-2059},
 	shorttitle = {{ACT}},
@@ -16,7 +16,7 @@
 }
 
 @article{artemis,
-	title = {Artemis: an integrated platform for visualization and analysis of high-throughput sequence-based experimental data},
+	title = {Artemis: {an} integrated platform for visualization and analysis of high-throughput sequence-based experimental data},
 	volume = {28},
 	copyright = {http://creativecommons.org/licenses/by-nc/3.0},
 	issn = {1367-4811, 1367-4803},
@@ -63,7 +63,7 @@
 	url = {https://journals.asm.org/doi/10.1128/IAI.71.5.2775-2786.2003},
 	doi = {10.1128/IAI.71.5.2775-2786.2003},
 	abstract = {ABSTRACT
-            
+
               We determined the complete genome sequence of
               Shigella flexneri
               serotype 2a strain 2457T (4,599,354 bp).
@@ -96,7 +96,7 @@
 }
 
 @article{k12,
-	title = {Escherichia coli {K}-12: a cooperatively developed annotation snapshot--2005},
+	title = {Escherichia coli {K}-12: {a} cooperatively developed annotation snapshot--2005},
 	volume = {34},
 	issn = {0305-1048, 1362-4962},
 	shorttitle = {Escherichia coli {K}-12},
@@ -113,22 +113,22 @@
 }
 
 @article{blast,
-	title = {{BLAST}+: architecture and applications},
+	title = {{BLAST}+: {architecture} and applications},
 	volume = {10},
 	issn = {1471-2105},
 	shorttitle = {{BLAST}+},
 	url = {https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-10-421},
 	doi = {10.1186/1471-2105-10-421},
 	abstract = {Abstract
-            
+
               Background
               Sequence similarity searching is a very important bioinformatics task. While Basic Local Alignment Search Tool (BLAST) outperforms exact methods through its use of heuristics, the speed of the current BLAST software is suboptimal for very long queries or database sequences. There are also some shortcomings in the user-interface of the current command-line applications.
-            
-            
+
+
               Results
               We describe features and improvements of rewritten BLAST software and introduce new command-line applications. Long query sequences are broken into chunks for processing, in some cases leading to dramatically shorter run times. For long database sequences, it is possible to retrieve only the relevant parts of the sequence, reducing CPU time and memory usage for searches of short queries against databases of contigs or chromosomes. The program can now retrieve masking information for database sequences from the BLAST databases. A new modular software library can now access subject sequence data from arbitrary data sources. We introduce several new features, including strategy files that allow a user to save and reuse their favorite set of options. The strategy files can be uploaded to and downloaded from the NCBI BLAST web site.
-            
-            
+
+
               Conclusion
               The new BLAST command-line applications, compared to the current BLAST tools, demonstrate substantial speed improvements for long queries as well as chromosome length database sequences. We have also improved the user interface of the command-line applications.},
 	language = {en},
