### 2025-12-17 #9004 @martinghunt
@abhishektiwari - thanks, I just made all the changes you listed. And thanks for the tip with `{}`, that made it easy to fix the reference titles.
