### 2025-12-19 #9004 @martinghunt
I've changed the text in the figshare entry to say "v1.1.0". Sorry but there's nothing I can do about figshare adding the "version 3" - it does indeed relate to upload and not the software version.

I've also edited to say United Kingdom everywhere instead of UK (in the paper and in figshare text).
