### 2026-01-30 0b02f27
Add JOSS badge/citation
